requirements.txt
paramiko==3.3.1
customtkinter==5.2.1
Pillow==10.1.0
flask==3.0.0
